Flytt felles mocking inn i egen klasse
